+++ src/static/util/dbfs.js
@@ -4,7 +4,7 @@
  * @author aKuad
  */
 
-import { typeof_detail } from "./typeof_detail.js";
+import { typeof_detail } from "../typeof_detail.js";
 
 
 /**
