+++ src/tests/js_denotest/AudioMixer.test.ts
@@ -15,7 +15,7 @@ import { AudioMixer, MaxLanesReachedError, NonExistingLaneIdError } from "../../
 import { ONE_FRAME_SAMPLES } from "../../static/AUDIO_PARAM.js";
 import { LaneInfo } from "../../static/packet_conv/LaneInfo.js";
 import { LaneLoudness } from "../../static/packet_conv/LaneLoudness.js";
-import { dbfs_float } from "../../static/dbfs.js";
+import { dbfs_float } from "../../static/util/dbfs.js";
 
 
 Deno.test(async function true_cases(t) {
